src/main/java/com/iteratrlearning/shu_book/chapter_02/BankStatementCSVParser.java
package com.iteratrlearning.shu_book.chapter_02;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.util.List;

import static java.util.stream.Collectors.toList;

public class BankStatementCSVParser implements BankStatementParser {

    private static final DateTimeFormatter DATE_PATTERN = DateTimeFormatter.ofPattern("dd-MM-yyy");

    public BankStatement parseFrom(String line) {
        String[] columns = line.split(",");

        LocalDate date = LocalDate.parse(columns[0], DATE_PATTERN);
        double amount = Double.parseDouble(columns[1]);

        return new BankStatement(date, amount, columns[2]);
    }

    public List<BankStatement> parseLinesFrom(List<String> lines) {
        return lines.stream().map(this::parseFrom).collect(toList());
    }


    // TODO: interface extractAmount(), extractDate(), extractDescription(), parseFromLine()?
    // CSV BankStatement

}
